requirements.txt
google-cloud
pandas
seaborn
matplotlib
